refactor: CodeGen.js -> generate/code-gen.js
