+++ packages/compile/src/index.js
@@ -2,7 +2,7 @@
 
 export { canonicalName } from './_shared/helpers.js'
 export { readDat } from './_shared/read-dat.js'
-export { generateCode } from './CodeGen.js'
+export { generateCode } from './generate/code-gen.js'
 export { parseAndGenerate, printNames } from './Entry.js'
 export { parseModel } from './Parser.js'
 export { preprocessModel } from './preprocess/preprocessor.js'
